Organized data Extant / Paleo
Files separated for  extant and paleo data. some files changed to csv format and were included new information (ecoregion code) to improve the script

It was easier to built a neu foldar than  with separated scripts  for each analysis.
The working directory should be set in the folder : "GABI-biogeography" so it has acces to all data

requires to install the folowing packages:
RODBC (for Windows)
Hmisc ( for Mac/Linux)
raster